wip: calculate amount

+++ money-calculator-back/src/main/java/com/moneycalculator/back/services/TransactionService.java
@@ -20,5 +20,5 @@ public interface TransactionService {
 
     TransactionIdTypeTotalDTO deleteTransaction(Integer id);
 
-    BigDecimal calculateTotalAmount(List<Transaction> transactions);
+    Double calculateTotalAmount(List<Transaction> transactions);
 }
